DSTNetPlus/requirements.txt
addict
future
lmdb
numpy>=1.17
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
torch>=1.7
torchvision
tqdm
yapf
